requirements.txt
opencv-python
supervision
inference
tkinter
Pillow
streamlit